requirements.txt
rpi_ws281x
pytz~=2022.6
websockets~=10.4
ADS1x15-ADC